chore(reorg): move option-set to new file
