+++ src/data-workspace/inputs/index.js
@@ -3,3 +3,4 @@ export * from './long-text.js'
 export * from './inputs.js'
 export * from './true-only-checkbox.js'
 export * from './boolean-radios.js'
+export * from './option-set.js'
